create event route fix and user profile linj\k
