+++ .kilo/kilo.jsonc
@@ -1,0 +1,4 @@
+{
+  "$schema": "https://app.kilo.ai/config.json",
+  "snapshot": false
+}
