+++ apps/web/DOCS/COMPONENTS.md
@@ -90,30 +90,46 @@ import { FormField } from "@/components/ui/form-field";
 />
 ```
 
-### `components/ui/EmptyState.tsx`
+### `components/ui/empty-state.tsx`
 
 Purpose:
-- Displayed when no content (e.g., events) matches the active filters. Accepts a title, message, and an optional call-to-action (CTA).
+- Displayed when no content (e.g., events) matches the active filters. Accepts a title, description, and an optional call-to-action (CTA).
 
 **Props**
 | Name | Type | Required | Default | Description |
 |---|---|---|---|---|
 | `title` | `string` | Yes | - | Heading text |
-| `message` | `string` | Yes | - | Supporting message / body copy |
-| `ctaLabel` | `string` | No | - | Label for the optional call-to-action button |
-| `ctaLink` | `string` | No | - | Where the CTA button navigates to |
+| `description` | `string` | Yes | - | Supporting message / body copy |
+| `icon` | `React.ReactNode` | No | - | Icon to display (e.g. `<img>`, `<Image />`, or any SVG element) |
+| `action` | `{ label: string; onClick?: () => void; href?: string }` | No | - | Optional call-to-action; renders as `<Link>` when `href` is provided, or as a `<Button>` when `onClick` is provided |
 | `illustrationSrc` | `string` | No | `"/icons/404-illustration.svg"` | Override the default illustration src |
 
 **Usage Example:**
 
 ```tsx
-import { EmptyState } from "@/components/ui/EmptyState";
+import { EmptyState } from "@/components/ui/empty-state";
+import Image from "next/image";
 
+// With icon and click handler
+<EmptyState
+  icon={
+    <Image
+      src="/icons/search.svg"
+      alt="Search"
+      width={32}
+      height={32}
+    />
+  }
+  title="No events found"
+  description="Try adjusting your filters to find what you're looking for."
+  action={{ label: "Clear Search", onClick: () => setSearch("") }}
+/>
+
+// With navigation link
 <EmptyState
   title="No events found"
-  message="Try adjusting your filters or search criteria."
-  ctaLabel="Create Event"
-  ctaLink="/events/new"
+  description="There are no events matching your current filters."
+  action={{ label: "Create an Event", href: "/events/create" }}
 />
 ```
